fix bug that the min and max elem in  function middleSplit_ is  wrong
if  span is always equal to max_span and equal to 0,if condition(span > (1 - EPS) * max_span) is always false, the  min_elem and max_elem are always the default initailized value 0, this is a bug. so the max_elem and min_elem should inlitialized while setting max_span.
This bug could be occur while the dataset store points which all have the same location(x,y,z).